Probabilistic Data Synthesis
- Changed probability analysis to probability data synthesis
- Added probability analysis file
- Added documentation for probability data analysis
